tectonic/core-strutils.h: add this
Pulling more functions out of internals.h, which I'd like to get rid of since
its name is not super helpful and it currently contains a big mish-mash of
functionality.
